+++ lib/page_magic/element/selector.rb
@@ -1,21 +1,9 @@
 # frozen_string_literal: true
-
+require 'page_magic/element/selector/model'
 module PageMagic
   class Element # Capybara::Finder
     # class Selector - models the selection criteria understood by Capybara
     class Selector
-      class Instance
-        attr_reader :args, :options
-        def initialize(args, options={})
-          @args = args
-          @options = options
-        end
-
-        def ==(other)
-          other.args == self.args && other.options == self.options
-        end
-      end
-
       class << self
         # Find a Selector using it's name
         # @param [Symbol] name the name of the required Selector in snakecase format. See class constants for available
@@ -48,7 +36,7 @@ module PageMagic
       # @param [Hash] locator the selection method and its parameter. E.g. text: 'click me'
       def build(element_type, locator, options:{})
         array = [type(element_type), selector, formatter.call(locator)].compact
-        Instance.new(array, self.options.merge(options))
+        Model.new(array, self.options.merge(options))
       end
 
       private
