+++ lib/page_magic/element/selector/methods.rb
@@ -1,0 +1,16 @@
+module PageMagic
+  class Element
+    class Selector
+      # module SelectorMethods - adds method for getting and setting an element selector
+      module Methods
+        # Gets/Sets a selector
+        # @param [Hash<symbol,string>] selector method for locating the browser element. E.g. text: 'the text'
+        def selector(selector = nil)
+          return @selector unless selector
+
+          @selector = selector
+        end
+      end
+    end
+  end
+end
