+++ lib/page_magic/element/selector/model.rb
@@ -1,0 +1,11 @@
+class Model
+  attr_reader :args, :options
+  def initialize(args, options={})
+    @args = args
+    @options = options
+  end
+
+  def ==(other)
+    other.args == self.args && other.options == self.options
+  end
+end
